Move common tests into its own test module
